dehack example
inheratance in C++ API extending Js classes is not supported.

The previous implementation was a hacky way and resulted in errors
after an upgrade. It has now been removed from the example and the
correct way is not to use inheratance in these classes. They are
only supported as "driver" subclasses allowing to inject Js logic
in callbacks from C++ and the actual logic implementation should
reside in the collaborator objects (composition vs inheratance) of
the driver.