fix cpgf class registration to module
when a nested class is created it is first registered
incorrectly (as it does not have the parent assigned yet).

Once the new parent is set, the original qualified name
was cached - this disabled usage of e.g. container iterators
as e.g. vector<a>::iterator would be overriden by vector<b>::iterator
(both being counted as only "iterator").

This patch resets the qualified names and forces to
re-register the class into the module with the new fqn.